llm_microservice/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.10.0
pypdf==3.17.4
python-dotenv==1.0.0
pydantic==2.5.3
streamlit==1.31.0
